+++ core/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 ### Changed
 
 - `trackSession`: Callback receives an authenticated fetch as second parameter when logged in
+- `observeSession`: Returns an rxjs observable that allows to observe the current session state and all changes to it
 
 ## 0.11.0